requirements.txt
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
colorzero==2.0
cryptography==46.0.7
deprecation==2.1.0
et-xmlfile==2.0.0
exceptiongroup==1.3.1
fsspec==2025.10.0
gpiozero==2.0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib-metadata==4.13.0
importlib-resources==5.13.0
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
openpyxl==3.1.5
packaging==26.0
panda==0.3.1
postgrest==2.28.3
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.20.0
pyiceberg==0.10.0
pyjwt==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.28.3
requests==2.32.5
rich==14.3.3
six==1.17.0
sortedcontainers==2.4.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
